Changed all decimals to doubles in probabilities, fixed bug in previous commit

+++ src/MineDotNet/AI/Solvers/Border.cs
@@ -8,14 +8,14 @@ namespace MineDotNet.AI.Solvers
         public Border(IList<Cell> cells)
         {
             Cells = cells;
-            Probabilities = new Dictionary<Coordinate, decimal>();
+            Probabilities = new Dictionary<Coordinate, double>();
             Verdicts = new Dictionary<Coordinate, bool>();
         }
         public IList<Cell> Cells { get; set; }
         public IList<IDictionary<Coordinate, bool>> ValidCombinations { get; set; }
         public int MinMineCount { get; set; }
         public int MaxMineCount { get; set; }
-        public IDictionary<Coordinate, decimal> Probabilities { get; set; }
+        public IDictionary<Coordinate, double> Probabilities { get; set; }
         public IDictionary<Coordinate, bool> Verdicts { get; set; }
         public bool SolvedFully { get; set; }
     }
