+++ src/MineDotNet/Etc/Maths.cs
@@ -7,7 +7,7 @@ namespace MineDotNet.Etc
 {
     public static class Maths
     {
-        private static IDictionary<int, IDictionary<int, decimal>> CombinationRatios { get; set; }
+        private static IDictionary<int, IDictionary<int, double>> CombinationRatios { get; set; }
 
         static Maths()
         {
@@ -16,21 +16,21 @@ namespace MineDotNet.Etc
 
         private static void InitializeCombinationRatios()
         {
-            CombinationRatios = new Dictionary<int, IDictionary<int, decimal>>();
+            CombinationRatios = new Dictionary<int, IDictionary<int, double>>();
             for (var n = 0; n < 1000; n++)
             {
-                var ratios = new Dictionary<int, decimal>();
+                var ratios = new Dictionary<int, double>();
                 CombinationRatios[n] = ratios;
                 ratios[0] = 1;
                 for (var k = 1; k <= n; k++)
                 {
-                    var ratio = (n + 1 - k) / (decimal)k;
+                    var ratio = (n + 1 - k) / (double)k;
                     ratios[k] = ratio;
                 }
             }
         }
 
-        public static decimal CombinationRatio(int from, int count)
+        public static double CombinationRatio(int from, int count)
         {
             if (count > from)
             {
